Clients: Fix session-expiry-interval error parsing

+++ client/client_shared.h
@@ -119,7 +119,7 @@ struct mosq_config {
 	bool pretty; /* sub, rr */
 	unsigned int timeout; /* sub */
 	int sub_opts; /* sub */
-	long session_expiry_interval;
+	long long session_expiry_interval;
 	int random_filter; /* sub */
 	int transport;
 #ifndef WIN32
